LayoutTests/imported/w3c/web-platform-tests/xhr/overridemimetype-blob-expected.txt


FAIL Use text/xml as fallback MIME type assert_equals: expected "text/xml" but got ""
PASS Use text/xml as fallback MIME type, 2 
FAIL Loading data… promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function (near '...tests.forEach...')"
